Rename shared module
