+++ src/main/kotlin/com/projectcitybuild/features/teleporthistory/listeners/TeleportListener.kt
@@ -4,7 +4,7 @@ import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.entities.CrossServerLocation
 import com.projectcitybuild.features.teleporting.events.PlayerPreSummonEvent
 import com.projectcitybuild.features.teleporting.events.PlayerPreTeleportEvent
-import com.projectcitybuild.integrations.shared.crossteleport.events.PlayerPreLocationTeleportEvent
+import com.projectcitybuild.integrations.internal.crossteleport.events.PlayerPreLocationTeleportEvent
 import com.projectcitybuild.modules.config.ConfigKey
 import com.projectcitybuild.modules.config.PlatformConfig
 import com.projectcitybuild.repositories.LastKnownLocationRepositoy
