+++ src/main/kotlin/com/projectcitybuild/features/teleporthistory/usecases/BackUseCase.kt
@@ -3,7 +3,7 @@ package com.projectcitybuild.features.teleporthistory.usecases
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.core.utilities.Result
 import com.projectcitybuild.core.utilities.Success
-import com.projectcitybuild.integrations.shared.crossteleport.LocationTeleporter
+import com.projectcitybuild.integrations.internal.crossteleport.LocationTeleporter
 import com.projectcitybuild.repositories.LastKnownLocationRepositoy
 import org.bukkit.entity.Player
 import javax.inject.Inject
