+++ src/main/kotlin/com/projectcitybuild/integrations/internal/crossteleport/CrossServerTeleportModule.kt
@@ -1,8 +1,8 @@
-package com.projectcitybuild.integrations.shared.crossteleport
+package com.projectcitybuild.integrations.internal.crossteleport
 
 import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.core.contracts.SpigotFeatureModule
-import com.projectcitybuild.integrations.shared.crossteleport.listeners.TeleportOnJoinListener
+import com.projectcitybuild.integrations.internal.crossteleport.listeners.TeleportOnJoinListener
 import javax.inject.Inject
 
 class CrossServerTeleportModule @Inject constructor(
