+++ src/main/kotlin/com/projectcitybuild/integrations/internal/crossteleport/CrossServerTeleportQueue.kt
@@ -1,4 +1,4 @@
-package com.projectcitybuild.integrations.shared.crossteleport
+package com.projectcitybuild.integrations.internal.crossteleport
 
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.core.utilities.Result
