+++ src/main/kotlin/com/projectcitybuild/integrations/internal/crossteleport/listeners/TeleportOnJoinListener.kt
@@ -1,9 +1,9 @@
-package com.projectcitybuild.integrations.shared.crossteleport.listeners
+package com.projectcitybuild.integrations.internal.crossteleport.listeners
 
 import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.core.utilities.Success
-import com.projectcitybuild.integrations.shared.crossteleport.CrossServerTeleportQueue
+import com.projectcitybuild.integrations.internal.crossteleport.CrossServerTeleportQueue
 import com.projectcitybuild.modules.logger.PlatformLogger
 import com.projectcitybuild.modules.textcomponentbuilder.send
 import org.bukkit.event.EventHandler
