+++ src/test/com/projectcitybuild/integrations/internal/crossteleport/LocationTeleporterTest.kt
@@ -1,10 +1,10 @@
-package com.projectcitybuild.integrations.shared.crossteleport
+package com.projectcitybuild.integrations.internal.crossteleport
 
 import com.projectcitybuild.stubs.CrossServerLocationMock
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.core.utilities.Success
 import com.projectcitybuild.entities.SubChannel
-import com.projectcitybuild.integrations.shared.crossteleport.events.PlayerPreLocationTeleportEvent
+import com.projectcitybuild.integrations.internal.crossteleport.events.PlayerPreLocationTeleportEvent
 import com.projectcitybuild.modules.channels.ProxyMessenger
 import com.projectcitybuild.modules.config.ConfigKey
 import com.projectcitybuild.modules.config.PlatformConfig
